Knocked through cowboy to webrick!
So, the problem is the API resides in heroku. It has two layers of HTTP service.

The first is "cowboy", it should route the request from the shared public IP to the specific
application somewhere within heroku based on "host" field.

The second is webrick, which is a ruby web sever.

So, basically, with a lot of effort in reverse engineering tcpdump binary log, I managed to
deliver request body through to webrick, which says it's bad. Well, better than nothing, webrick!